+++ src/vue/forms/DepositForm.vue
@@ -12,12 +12,19 @@
           <div class="app__form-row">
             <div class="app__form-field">
               <select-field
-                v-model="selectedAsset"
-                :values="assets"
-                key-as-value-text="nameAndCode"
+                :value="selectedAsset"
+                @input="setAssetByCode"
                 :label="'deposit-form.asset' | globalize"
                 :disabled="formMixin.isDisabled"
-              />
+              >
+                <option
+                  v-for="asset in assets"
+                  :key="asset.code"
+                  :value="asset.code"
+                >
+                  {{ asset.nameAndCode }}
+                </option>
+              </select-field>
             </div>
           </div>
         </div>
@@ -92,6 +99,10 @@ export default {
   },
   mixins: [FormMixin],
 
+  props: {
+    assetCode: { type: String, default: '' },
+  },
+
   data () {
     return {
       CoinpaymentsDepositModule,
@@ -122,7 +133,8 @@ export default {
         .filter(item => item.isDepositable)
 
       if (this.assets.length) {
-        this.selectedAsset = this.assets[0]
+        this.selectedAsset = this.assets.find(a => a.code === this.assetCode) ||
+          this.assets[0]
       }
       this.isLoaded = true
     } catch (e) {
@@ -135,6 +147,10 @@ export default {
     ...mapActions({
       loadBalances: vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS,
     }),
+
+    setAssetByCode (code) {
+      this.selectedAsset = this.assets.find(item => item.code === code) || {}
+    },
   },
 }
 </script>