+++ src/vue/pages/Fees.vue
@@ -16,11 +16,18 @@
             />
             <select-field
               v-if="isLoaded && assets.length"
-              v-model="asset"
-              :values="assets"
-              key-as-value-text="nameAndCode"
+              :value="asset.code"
+              @input="setAssetByCode"
               class="app__select app__select--no-border"
-            />
+            >
+              <option
+                v-for="asset in assets"
+                :key="asset.code"
+                :value="asset.code"
+              >
+                {{ asset.nameAndCode }}
+              </option>
+            </select-field>
             <skeleton-loader
               v-else
               template="bigString"
@@ -113,6 +120,10 @@ export default {
     ...mapActions({
       loadBalances: vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS,
     }),
+
+    setAssetByCode (code) {
+      this.asset = this.assets.find(item => item.code === code)
+    },
 
     async initAssetSelector () {
       await this.loadAssets()