+++ CHANGELOG.md
@@ -27,6 +27,13 @@ for further information about branching and tagging conventions.
 - Cap asset field to the create sale form
 - Display amount + fee in movements list
 - Added skeleton loader template by elements
+- "custom-select" package for displaying customized select field
+- Max button to amount input fields
+- "compare" & "format" methods to `MathUtil` class
+- Name attributes to 2FA form inputs
+- Emitted value normalizers for input field:
+  - "normalizeRange"
+  - "normalizeDecimalPrecision"
 
 #### Changed
 - Now loading account converted balances in
@@ -52,13 +59,19 @@ for further information about branching and tagging conventions.
   verification form
 - Now using "d3Format" and SI prefixes while formatting chart percent hints
 - Replaced spinner loader on skeleton loader in almost all templates
+- Now displaying asset for incoming withdrawal requests
+- Now passing values to select field using a slot and "option" tags
+- Now using `MathUtil.compare` method in validators
+- Now using `MathUtil.format` for formatting money, numbers, and percents
 
 #### Removed
 - Cursor pointer on disabled select
 - Price history chart from sale overview page
 - Validation for address line 2
 - Fixed fee from invest form
-- Invest drawer from sale stae widget & opportunity details components
+- Invest drawer from sale state widget & opportunity details components
+- "numeral" package
+- Order number format & filter
 
 #### Fixed
 - Don't show title in kyc status message, if no additional external information
@@ -70,6 +83,8 @@ for further information about branching and tagging conventions.
 - Disabling invest form while submitting
 - Different borders with attribute of readonly on the field
 - Close transfer drawer on Dashboard after submitting form
+- Displaying selected asset in drawers on movements page
+- Translations on sale details drawer
 
 ## [1.7.0-rc.2] - 2019-05-21
 ### "Under the hood" changes